+++ locals-naming.tf
@@ -2,5 +2,5 @@ locals {
   name_prefix = lower(var.name_prefix)
   name_suffix = lower(var.name_suffix)
 
-  vwan_name = coalesce(var.custom_vwan_name, azurecaf_name.virtual_wan_caf.result)
+  vwan_name = coalesce(var.custom_vwan_name, data.azurecaf_name.virtual_wan_caf.result)
 }