+++ modules/firewall/locals-naming.tf
@@ -1,5 +1,5 @@
 locals {
   name_prefix = lower(var.name_prefix)
   name_suffix = lower(var.name_suffix)
-  fw_name     = coalesce(var.custom_name, azurecaf_name.azure_firewall_caf.result)
+  fw_name     = coalesce(var.custom_name, data.azurecaf_name.azure_firewall_caf.result)
 }