+++ modules/virtual-hub/versions.tf
@@ -1,9 +1,13 @@
 terraform {
-  required_version = "~> 1.0"
+  required_version = "~> 1.3"
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = "~> 3.0"
+      version = "~> 3.22"
+    }
+    azurecaf = {
+      source  = "aztfmod/azurecaf"
+      version = "~> 1.2, >= 1.2.22"
     }
   }
 }